Memory/requirements.txt
# robot_memory — Python dependencies
# Install with:  pip install -r requirements.txt

# ── Database ──────────────────────────────────────────────────────────────────
asyncpg>=0.29.0          # async PostgreSQL driver

# ── Environment ───────────────────────────────────────────────────────────────
python-dotenv>=1.0.0     # loads .env into os.environ at startup

# ── Vector / embeddings ───────────────────────────────────────────────────────
pgvector>=0.2.4          # Python client for pgvector (asyncpg codec registration)

# ── LLM clients (consolidator.py) ────────────────────────────────────────────
# openai>=1.30 provides the AsyncOpenAI client used for ALL providers:
#   - Groq    (llama-3.3-70b-versatile)   → ROBOT_LLM_PROVIDER=groq
#   - Together AI (Llama-3-70b)           → ROBOT_LLM_PROVIDER=together
#   - Ollama local                        → ROBOT_LLM_PROVIDER=ollama
# All use the OpenAI-compatible /v1/chat/completions endpoint.
openai>=1.30.0

# ── Pathfinding fallback ──────────────────────────────────────────────────────
# networkx is only needed if you replace the built-in Dijkstra with networkx.
# The built-in implementation has no extra dependencies.
# networkx>=3.2

# ── Utilities ─────────────────────────────────────────────────────────────────
numpy>=1.26              # embedding arithmetic helpers (optional)

# ── Web API / server ──────────────────────────────────────────────────────────
fastapi>=0.115.0         # API framework
uvicorn>=0.30.0          # ASGI server



# ── Sensor I/O ────────────────────────────────────────────────────────────────
opencv-python>=4.9.0        # camera capture + frame encoding
sounddevice>=0.4.6          # cross-platform microphone capture (PortAudio)
 
# ── Speech-to-text  (choose one) ──────────────────────────────────────────────
# Option A — Local Whisper (recommended, no API key, runs on CPU)
faster-whisper>=1.0.0
 
 
# ── HTTP / geolocation ────────────────────────────────────────────────────────
requests>=2.31.0            # IP geolocation via ipapi.co